Feat: NightTime CW testcases
