+++ py_src/Night_CW_Algorithm/testcases/decrease.json
@@ -1,0 +1,440 @@
+[
+  {
+    "target": "BakerLab.CW.FP/TONS",
+    "datapoints": [
+      [
+        17.019910416,
+        1706832000000
+      ],
+      [
+        16.968650537,
+        1706918400000
+      ],
+      [
+        15.66481875,
+        1707004800000
+      ],
+      [
+        15.464358333,
+        1707091200000
+      ],
+      [
+        16.111213541,
+        1707177600000
+      ],
+      [
+        15.769204166,
+        1707264000000
+      ],
+      [
+        14.2420125,
+        1707350400000
+      ],
+      [
+        13.691845833,
+        1707436800000
+      ],
+      [
+        13.638927083,
+        1707523200000
+      ],
+      [
+        13.8560375,
+        1707609600000
+      ],
+      [
+        13.469183157,
+        1707696000000
+      ],
+      [
+        12.494123958,
+        1707782400000
+      ],
+      [
+        15.92663125,
+        1707868800000
+      ],
+      [
+        19.937178494,
+        1707955200000
+      ],
+      [
+        19.393288541,
+        1708041600000
+      ],
+      [
+        16.97245,
+        1708128000000
+      ],
+      [
+        19.7352625,
+        1708214400000
+      ],
+      [
+        16.684215625,
+        1708300800000
+      ],
+      [
+        16.913818947,
+        1708387200000
+      ],
+      [
+        18.726039583,
+        1708473600000
+      ],
+      [
+        18.902923958,
+        1708560000000
+      ],
+      [
+        19.116726041,
+        1708646400000
+      ],
+      [
+        18.659797916,
+        1708732800000
+      ],
+      [
+        16.928683333,
+        1708819200000
+      ],
+      [
+        17.711631578,
+        1708905600000
+      ],
+      [
+        22.961672916,
+        1708992000000
+      ],
+      [
+        7.693141052,
+        1709078400000
+      ],
+      [
+        6.607101875,
+        1709164800000
+      ],
+      [
+        21.437802083,
+        1709251200000
+      ],
+      [
+        20.554098958,
+        1709337600000
+      ],
+      [
+        20.332334375,
+        1709424000000
+      ],
+      [
+        41.118458333,
+        1709510400000
+      ],
+      [
+        23.066156842,
+        1709596800000
+      ],
+      [
+        21.061771875,
+        1709683200000
+      ],
+      [
+        21.164628125,
+        1709769600000
+      ],
+      [
+        21.250411458,
+        1709856000000
+      ],
+      [
+        21.275952083,
+        1709942400000
+      ],
+      [
+        21.072354166,
+        1710028800000
+      ],
+      [
+        20.705432291,
+        1710115200000
+      ],
+      [
+        20.647673958,
+        1710201600000
+      ],
+      [
+        48.167578125,
+        1710288000000
+      ],
+      [
+        24.72353802,
+        1710374400000
+      ],
+      [
+        19.5609,
+        1710460800000
+      ],
+      [
+        18.911425,
+        1710547200000
+      ],
+      [
+        18.469716842,
+        1710633600000
+      ],
+      [
+        18.201178125,
+        1710720000000
+      ],
+      [
+        18.083319791,
+        1710806400000
+      ],
+      [
+        18.131917894,
+        1710892800000
+      ],
+      [
+        17.939251041,
+        1710979200000
+      ],
+      [
+        17.9987,
+        1711065600000
+      ],
+      [
+        18.429479166,
+        1711152000000
+      ],
+      [
+        19.429051041,
+        1711238400000
+      ],
+      [
+        19.609248958,
+        1711324800000
+      ],
+      [
+        19.747482291,
+        1711411200000
+      ],
+      [
+        20.568504375,
+        1711497600000
+      ],
+      [
+        20.740479166,
+        1711584000000
+      ],
+      [
+        19.116526315,
+        1711670400000
+      ],
+      [
+        19.153413541,
+        1711756800000
+      ],
+      [
+        19.0289625,
+        1711843200000
+      ],
+      [
+        18.576096875,
+        1711929600000
+      ],
+      [
+        18.090727083,
+        1712016000000
+      ],
+      [
+        18.149693548,
+        1712102400000
+      ],
+      [
+        17.848409375,
+        1712188800000
+      ],
+      [
+        18.697736363,
+        1712275200000
+      ],
+      [
+        15.71010909,
+        1712534400000
+      ],
+      [
+        21.431797297,
+        1713398400000
+      ],
+      [
+        20.109583333,
+        1713484800000
+      ],
+      [
+        19.844670526,
+        1713571200000
+      ],
+      [
+        19.541388541,
+        1713657600000
+      ],
+      [
+        19.520716666,
+        1713744000000
+      ],
+      [
+        44.6266125,
+        1713830400000
+      ],
+      [
+        22.20259125,
+        1713916800000
+      ],
+      [
+        18.461589687,
+        1714003200000
+      ],
+      [
+        23.21778125,
+        1714089600000
+      ],
+      [
+        21.084109375,
+        1714176000000
+      ],
+      [
+        56.994178125,
+        1714262400000
+      ],
+      [
+        136.817998958,
+        1714348800000
+      ],
+      [
+        87.428302083,
+        1714435200000
+      ],
+      [
+        68.878770833,
+        1714521600000
+      ],
+      [
+        88.264705263,
+        1714608000000
+      ],
+      [
+        83.074406382,
+        1714694400000
+      ],
+      [
+        53.202539583,
+        1714780800000
+      ],
+      [
+        22.618726595,
+        1714867200000
+      ],
+      [
+        51.136347916,
+        1714953600000
+      ],
+      [
+        63.452336458,
+        1715040000000
+      ],
+      [
+        98.827767708,
+        1715126400000
+      ],
+      [
+        25.56043125,
+        1715212800000
+      ],
+      [
+        21.09725625,
+        1715299200000
+      ],
+      [
+        20.6730125,
+        1715385600000
+      ],
+      [
+        20.833546875,
+        1715472000000
+      ],
+      [
+        31.792903191,
+        1715558400000
+      ],
+      [
+        95.756947916,
+        1715644800000
+      ],
+      [
+        61.085838541,
+        1715731200000
+      ],
+      [
+        80.86438387,
+        1715817600000
+      ],
+      [
+        81.904962105,
+        1715904000000
+      ],
+      [
+        77.894366315,
+        1715990400000
+      ],
+      [
+        105.830327083,
+        1716076800000
+      ],
+      [
+        142.890035416,
+        1716163200000
+      ],
+      [
+        210.7585,
+        1716249600000
+      ],
+      [
+        223.255465625,
+        1716336000000
+      ],
+      [
+        154.732079479,
+        1716422400000
+      ],
+      [
+        111.739560416,
+        1716508800000
+      ],
+      [
+        96.740210752,
+        1716595200000
+      ],
+      [
+        121.268763541,
+        1716681600000
+      ],
+      [
+        198.472978494,
+        1716768000000
+      ],
+      [
+        96.80180625,
+        1716854400000
+      ],
+      [
+        35.007565833,
+        1716940800000
+      ],
+      [
+        19.288278125,
+        1717027200000
+      ]
+    ],
+    "metric": "Chilled Water"
+  }
+]
